pratica_2/requirements.txt
colorzero==2.0
gpiozero==2.0.1
RPi.GPIO==0.7.1; platform_machine == "armv7l" or platform_machine == "aarch64"
setuptools==80.9.0
